backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Dmitriy Golub <dg@mongo.one>"]

[tool.poetry.dependencies]
python = "^3.10"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
raven = "^6.10.0"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
typer = "^0.4.0"
progress = "^1.6"
pymotyc = "^0.9.0"
python-multipart = "^0.0.5"
emails = "^0.6"
gunicorn = "^20.1.0"
Jinja2 = "^3.1.2"
alembic = "^1.7.7"
aiohttp = "^3.8.1"
asyncio = "^3.4.3"
uvicorn = "^0.17.6"
motor = "^2.0.0"
cached-property = "^1.5.2"
celery = "^5.2.7"
importlib-metadata = "^4.11.4"
pydantic = {extras = ["email"], version = "^1.9.1"}
email-validator = "^1.2.1"
sqlalchemy-stubs = "^0.4"
platformdirs = "2.6.2"
fastapi = "^0.95.0"
httpx = "^0.23.3"
sentry-sdk = {extras = ["fastapi"], version = "^1.17.0"}
requests = "^2.28.2"
google-auth = "^2.16.3"
google-auth-oauthlib = "^1.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
bandit = "^1.7.5"
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.2"
flake8 = "^6.0.0"
autoflake = "^2.0.2"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
python-dotenv = "^1.0.0"
pytest-dotenv = "^0.5.2"
pre-commit = "^3.2.1"
pytest-env = "^0.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 160
profile = "black"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
python_classes = ["Test", "Describe"]
python_functions = ["test_", "it_", "and_", "but_", "they_"]
python_files = ["test_*.py",]
testpaths = ["app/tests",]
norecursedirs = [
    ".venv",
    "backend/app/app/tests"
]
